+++ packages/perseus-editor/src/multirenderer-editor.jsx
@@ -14,6 +14,7 @@ import {
     MultiItems,
 } from "@khanacademy/perseus";
 import {StyleSheet, css} from "aphrodite";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import lens from "hubble";
 import * as React from "react";
 import ReactDOM from "react-dom";