+++ packages/perseus/src/widgets/passage/passage-markdown.jsx
@@ -1,9 +1,9 @@
 /* eslint-disable react/sort-comp */
 // @flow
+import SimpleMarkdown from "@khanacademy/simple-markdown";
 import * as i18n from "@khanacademy/wonder-blocks-i18n";
 import PropTypes from "prop-types";
 import * as React from "react";
-import SimpleMarkdown from "simple-markdown";
 import _ from "underscore";
 
 const START_REF_PREFIX = "start-ref-";