+++ packages/perseus/src/widgets/unit.jsx
@@ -6,6 +6,7 @@
 // value?
 import * as KAS from "@khanacademy/kas";
 import * as i18n from "@khanacademy/wonder-blocks-i18n";
+// eslint-disable-next-line import/no-extraneous-dependencies
 import lens from "hubble";
 import $ from "jquery";
 import * as React from "react";
@@ -81,7 +82,7 @@ type DefaultProps = {|
  * shows and hides an error buddy. The error message is only shown after a
  * rolling two second delay, but hidden immediately on further typing.
  */
-class OldUnitInput extends React.Component<Props> {
+export class OldUnitInput extends React.Component<Props> {
     _errorTimeout: ?TimeoutID;
 
     static defaultProps: DefaultProps = {