+++ packages/perseus/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/perseus
 
+## 0.2.2
+
+### Patch Changes
+
+-   591420a6: Add @flow comment to math-input's index.js and missing props to ProvidedKeypad
+-   Updated dependencies [591420a6]
+    -   @khanacademy/math-input@0.5.1
+
+## 0.2.1
+
+### Patch Changes
+
+-   e56ab24b: Exclude all deps from bundle except for vendored deps
+-   Updated dependencies [ca76baa3]
+    -   @khanacademy/simple-markdown@0.8.1
+
 ## 0.2.0
 
 ### Minor Changes